mesh: move into shared place
